+++ pycbc/tmpltbank/coord_utils.py
@@ -19,7 +19,8 @@ import numpy
 from pycbc.tmpltbank.lambda_mapping import get_chirp_params
 from pycbc import conversions
 from pycbc import pnutils
-from pycbc.tmpltbank.em_progenitors import load_ns_sequence, remnant_masses
+from pycbc.neutron_stars import load_ns_sequence
+from pycbc.conversions import remnant_mass
 
 def estimate_mass_range(numPoints, massRangeParams, metricParams, fUpper,\
                         covary=True):
@@ -291,7 +292,7 @@ def get_random_mass(numPoints, massRangeParams):
             #     threshold mass]
             mask_bright_nsbh = numpy.zeros(len(mass1_nsbh), dtype=bool)
             if eta_nsbh.size != 0:
-                remnant = remnant_masses(eta_nsbh, mass2_nsbh, ns_sequence,
+                remnant = remnant_mass(eta_nsbh, mass2_nsbh, ns_sequence,
                                          spin1z_nsbh, 0.)
                 mask_bright_nsbh[remnant
                                  >
