+++ pycbc/tmpltbank/em_progenitors.py
@@ -16,12 +16,7 @@
 
 import math
 import numpy as np
-import pycbc
-import os.path
 import scipy.optimize
-import scipy.interpolate
-import logging
-import sys
 
 ##############################################################################
 # Innermost Stable Spherical Orbit (ISSO) solver in the Perez-Giz (PG)       #
@@ -172,189 +167,3 @@ def PG_ISSO_solver(chi,incl):
                                              args=(chi, incl))
 
     return solution
-
-##############################################################################
-# 2H 2-piecewise polytropic EOS, NS non-rotating equilibrium sequence        #
-# File format is: grav mass (Msun)   baryonic mass (Msun)    compactness     #
-#                                                                            #
-# Eventually, the function should call an NS sequence generator within LAL   #
-# the EOS prescribed by the user and store it.                               #
-##############################################################################
-def load_ns_sequence(eos_name):
-    """
-    Load the data of an NS non-rotating equilibrium sequence
-    generated using the equation of state (EOS) chosen by the
-    user.  [Only the 2H 2-piecewise polytropic EOS is currently
-    supported.  This yields NSs with large radiss (15-16km).]
-
-    Parameters
-    -----------
-    eos_name: string
-        NS equation of state label ('2H' is the only supported
-        choice at the moment)
-
-    Returns
-    ----------
-    ns_sequence: 3D-array
-        contains the sequence data in the form NS gravitational
-         mass (in solar masses), NS baryonic mass (in solar
-         masses), NS compactness (dimensionless)
-    max_ns_g_mass: float
-        the maximum NS gravitational mass (in solar masses) in
-        the sequence (this is the mass of the most massive stable
-        NS)
-    """
-    ns_sequence = []
-
-    if eos_name == '2H':
-        ns_sequence_path = os.path.join(pycbc.tmpltbank.NS_SEQUENCE_FILE_DIRECTORY, 'equil_2H.dat')
-        ns_sequence = np.loadtxt(ns_sequence_path)
-    else:
-        err_msg = "Only the 2H EOS is currently supported."
-        err_msg += "If you plan to use a different NS EOS,"
-        err_msg += "be sure not to filter too many templates!\n"
-        logging.error(err_msg)
-        sys.exit(1)
-        raise Exception('Unsupported EOS!')
-
-    max_ns_g_mass = max(ns_sequence[:,0])
-
-    return (ns_sequence, max_ns_g_mass)
-
-##############################################################################
-# Given an NS equilibrium sequence and gravitational mass (in Msun), return  #
-# the NS baryonic mass (in Msun).                                            #
-##############################################################################
-def ns_g_mass_to_ns_b_mass(ns_g_mass, ns_sequence):
-    """
-    Determines the baryonic mass of an NS given its gravitational
-    mass and an NS equilibrium sequence.
-
-    Parameters
-    -----------
-    ns_g_mass: float
-        NS gravitational mass (in solar masses)
-    ns_sequence: 3D-array
-        contains the sequence data in the form NS gravitational
-         mass (in solar masses), NS baryonic mass (in solar
-         masses), NS compactness (dimensionless)
-
-    Returns
-    ----------
-    float
-        The NS baryonic mass (in solar massesr**3*(r**2*(r-6)+chi**2*(3*r+4))+
-        chi**4*(3*r*(r-2)+chi**2))
-    """
-    x = ns_sequence[:,0]
-    y = ns_sequence[:,1]
-    f = scipy.interpolate.interp1d(x, y)
-
-    return f(ns_g_mass)
-
-##############################################################################
-# Given an NS equilibrium sequence and gravitational mass (in Msun), return  #
-# the NS compactness.                                                        #
-##############################################################################
-def ns_g_mass_to_ns_compactness(ns_g_mass, ns_sequence):
-    """
-    Determines the compactness of an NS given its
-    gravitational mass and an NS equilibrium sequence.
-
-    Parameters
-    -----------
-    ns_g_mass: float
-        NS gravitational mass (in solar masses)
-    ns_sequence: 3D-array
-        contains the sequence data in the form NS gravitational
-         mass (in solar masses), NS baryonic mass (in solar
-         masses), NS compactness (dimensionless)
-
-    Returns
-    ----------
-    float
-        The NS compactness (dimensionless)
-    """
-    x = ns_sequence[:,0]
-    y = ns_sequence[:,2]
-    f = scipy.interpolate.interp1d(x, y)
-
-    return f(ns_g_mass)
-
-##############################################################################
-# NS-BH merger remnant mass [Foucart, Hinderer, Nissanke PRD 98, 081501(R)   #
-# (2018)].                                                                   #
-#                                                                            #
-# * THIS ASSUMES THE NS SPIN IS 0 (the user is warned about this).           #
-# * Physical parameters passed to remnant_mass (see sanity checks below)     #
-#   must not be unphysical.                                                  #
-# * Tilted BH spin cases are handled by using rISSO(chi,incl) where rISCO    #
-#   appears in the formula. This approximation was suggested in Stone, Loeb, #
-#   Berger, PRD 87, 084053 (2013) for the previous NS-BH remnant mass fit of #
-#   Foucart, PRD 86, 124007 (2012).                                          #
-##############################################################################
-def remnant_mass(eta, ns_g_mass, ns_sequence, chi, incl):
-    """
-    Function that determines the remnant disk mass of
-    an NS-BH system using the fit to numerical-relativity
-    results discussed in Foucart, Hinderer, Nissanke PRD 98, 081501(R) (2018).
-
-    Parameters
-    -----------
-    eta: float
-        the symmetric mass ratio of the binary
-    ns_g_mass: float
-        NS gravitational mass (in solar masses)
-    ns_sequence: 3D-array
-        contains the sequence data in the form NS gravitational
-         mass (in solar masses), NS baryonic mass (in solar
-         masses), NS compactness (dimensionless)
-    chi: float
-        the BH dimensionless spin parameter
-    incl: float
-        the inclination angle between the BH spin and the orbital
-        angular momentum in radians
-
-    Returns
-    ----------
-    remnant_mass: float
-        The remnant mass in solar masses
-    """
-
-    # Sanity checks on eta and chi
-    if not (eta>0. and eta<=0.25 and abs(chi)<=1 and ns_g_mass>0):
-        err_msg = "The BH spin magnitude must be <= 1,"
-        err_msg += "eta must be between 0 and 0.25,"
-        err_msg += "and the NS mass must be positive."
-        logging.error(err_msg)
-        info_msg = "The function remnant_mass was launched with"
-        info_msg += "ns_mass={0}, eta={1}, chi={2},"
-        info_msg += "inclination={3}\n'.format(ns_g_mass, eta, chi, incl)"
-        logging.info(info_msg)
-        sys.exit(1)
-        raise Exception('Unphysical parameters!')
-
-    # NS compactness and rest mass
-    ns_compactness = ns_g_mass_to_ns_compactness(ns_g_mass, ns_sequence)
-    ns_b_mass = ns_g_mass_to_ns_b_mass(ns_g_mass, ns_sequence)
-
-    # Fit parameters and tidal correction
-    alpha = 0.406
-    beta  = 0.139
-    gamma = 0.255
-    delta = 1.761
-    # The remnant mass over the NS rest mass
-    remnant_mass = (max(alpha/eta**(1./3.)*(1 - 2*ns_compactness)
-                        - beta*ns_compactness / eta*PG_ISSO_solver(chi, incl)
-                        + gamma, 0.)) ** delta
-    # Convert to solar masses
-    remnant_mass = remnant_mass*ns_b_mass
-
-    return remnant_mass
-
-
-#############################################################################
-# Vectorized version of remnant_mass. The third argument (NS equilibrium    #
-# sequence) is excluded from vectorisation.                                 #
-#############################################################################
-remnant_masses = np.vectorize(remnant_mass)
-remnant_masses.excluded.add(2)
